% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoe_storage.R
\name{shoe_storage}
\alias{shoe_storage}
\title{Calculate capacity of a shoe storage}
\usage{
shoe_storage(shelf_capacity, number_shelves)
}
\arguments{
\item{shelf_capacity}{Define how many shoe pairs can fit on a shelf}

\item{number_shelves}{Define number of shelves}
}
\description{
Calculate capacity of a shoe storage
}
